% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{MFMouse}
\alias{MFMouse}
\title{MF GO-terms reduced GO-DAG for Mouse}
\format{
A list with 3 objects:
\describe{
  \item{v}{GO-terms in the respective nodes.}
  \item{dat.d}{Data frame object that shows how the respective nodes are connected.}
  \item{df}{Sketelon of the reduced GO-dags.}

}
}
\source{
<https://www.nature.com/articles/s41598-020-73326-3>
}
\usage{
MFMouse
}
\description{
A data set with list containing information about the reduced GO-BP DAG for Mouse.
}
\keyword{datasets}
