% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{Mouse}
\alias{Mouse}
\title{BP GO-terms as an Edgelist for Mouse}
\format{
A list containing 3 objects:
\describe{
  \item{BPlist}{BP GO-terms edgelist.}
  \item{MFlist}{MF GO-terms edgelist.}
  \item{CClist}{CC GO-terms edgelist.}
}
}
\source{
<https://www.nature.com/articles/s41598-020-73326-3>
}
\usage{
Mouse
}
\description{
A data set containing BP, MF and CC GO-terms as an edgelist for Mouse.
}
\keyword{datasets}
