% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{MouseAll}
\alias{MouseAll}
\title{All GO-terms for Mouse}
\format{
A data frame with 337505 rows and 2 variables:
\describe{
  \item{entrezgene_id}{The entrezgene_id.}
  \item{go_id}{The respective GO-terms.}
}
}
\source{
<https://www.nature.com/articles/s41598-020-73326-3>
}
\usage{
MouseAll
}
\description{
A data set with all the GO-terms for Mouse.
}
\keyword{datasets}
