% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{RatAll}
\alias{RatAll}
\title{All GO-terms for Rat}
\format{
A data frame with 262619 rows and 2 variables:
\describe{
  \item{entrezgene_id}{The entrezgene_id.}
  \item{go_id}{The respective GO-terms.}
}
}
\source{
<https://www.nature.com/articles/s41598-020-73326-3>
}
\usage{
RatAll
}
\description{
A data set with all the GO-terms for Rat.
}
\keyword{datasets}
